Add files via upload
Developed Netlogo and Thymio robotics projects as a part of the coursework for the course Artificial Life with Robotics. The robot followed other robots, walk-in flocks, detected the obstacles or lines and avoided them.

I divided the detection zones into three different layers far sight, near sight, and perspective space. By changing the values of these variables, the robots change their behavior like following other robots or a walk-in flock of robots.